Module log_store::kafka::util::record

Structs

Record 🔒

The minimal storage unit in the Kafka log store.

RecordMeta

The metadata of a record.

Enums

RecordType

The type of a record.

Constants

ESTIMATED_META_SIZE 🔒

The estimated size in bytes of a serialized RecordMeta. A record is guaranteed to have sizeof(meta) + sizeof(data) <= max_batch_byte - ESTIMATED_META_SIZE.

VERSION 🔒

The current version of Record.

Functions

convert_to_kafka_records 🔒

convert_to_multiple_entry 🔒

convert_to_naive_entry 🔒

maybe_emit_entry 🔒

For type of [Entry::Naive] Entry:

remaining_entries

Constructs entries from buffered_records