log_store::kafka

Module worker

Modules

dump_index ğŸ”’

flush ğŸ”’

produce ğŸ”’

update_high_watermark ğŸ”’

Structs

BackgroundProducerWorker ğŸ”’

DumpIndexRequest ğŸ”’

PendingRequest ğŸ”’

ProduceRequest ğŸ”’

ProduceResultHandle ğŸ”’

Receives the committed offsets when data has been committed to Kafka or an unrecoverable error has been encountered.

ProduceResultReceiver ğŸ”’

TruncateIndexRequest ğŸ”’

Enums

WorkerRequest ğŸ”’