log_store
Module metrics

Structs

METRIC_KAFKA_APPEND_BATCH_BYTES_TOTAL

Counter of bytes of the append_batch operation on the kafka logstore.

METRIC_KAFKA_APPEND_BATCH_ELAPSED

Timer of the append_batch operation on the kafka logstore.

METRIC_KAFKA_READ_BYTES_TOTAL

Counter of bytes of the read operation on the kafka logstore.

METRIC_KAFKA_READ_ELAPSED

Timer of the append_batch operation on the kafka logstore. This timer only measures the duration of the read operation, not measures the total duration of replay.

METRIC_LOGSTORE_OP_BYTES_TOTAL

Counters of bytes of each operation on a logstore.

METRIC_LOGSTORE_OP_ELAPSED

Timer of operations on a logstore.

METRIC_RAFT_ENGINE_APPEND_BATCH_BYTES_TOTAL

Counter of bytes of the append_batch operation on the raft-engine logstore.

METRIC_RAFT_ENGINE_APPEND_BATCH_ELAPSED

Timer of the append_batch operation on the raft-engine logstore.

METRIC_RAFT_ENGINE_READ_BYTES_TOTAL

Counter of bytes of the read operation on the raft-engine logstore.

METRIC_RAFT_ENGINE_READ_ELAPSED

Timer of the append_batch operation on the raft-engine logstore. This timer only measures the duration of the read operation, not measures the total duration of replay.

Constants

LOGSTORE_LABEL

Logstore label.

OPTYPE_LABEL

Operation type label.