Module log_store::metrics

Structs

METRIC_KAFKA_APPEND_BATCH_BYTES_TOTAL

Counter of bytes of the append_batch operation on the kafka logstore.

METRIC_KAFKA_APPEND_BATCH_CALLS_TOTAL

Counter of calls of the append_batch operation on the kafka logstore.

METRIC_KAFKA_APPEND_BATCH_ELAPSED

Timer of the append_batch operation on the kafka logstore.

METRIC_KAFKA_PRODUCE_RECORD_BYTES_TOTAL

Counter of bytes of the records produced by the kafka logstore.

METRIC_KAFKA_PRODUCE_RECORD_COUNTS

Counter of the numbers of the records produced by the kafka logstore.

METRIC_KAFKA_READ_BYTES_TOTAL

Counter of bytes of the read operation on the kafka logstore.

METRIC_KAFKA_READ_CALLS_TOTAL

Counter of calls of the read operation on the kafka logstore.

METRIC_KAFKA_READ_ELAPSED

Timer of the append_batch operation on the kafka logstore. This timer only measures the duration of the read operation, not measures the total duration of replay.

METRIC_KAFKA_READ_RECORD_BYTES_TOTAL

Counter of bytes of the records read by the kafka logstore.

METRIC_LOGSTORE_OP_BYTES_TOTAL

Counters of bytes of each operation on a logstore.

METRIC_LOGSTORE_OP_CALLS_TOTAL

Counters of calls of each operation on a logstore.

METRIC_LOGSTORE_OP_ELAPSED

Timer of operations on a logstore.

METRIC_RAFT_ENGINE_APPEND_BATCH_BYTES_TOTAL

Counter of bytes of the append_batch operation on the raft-engine logstore.

METRIC_RAFT_ENGINE_APPEND_BATCH_CALLS_TOTAL

Counter of calls of the append_batch operation on the raft-engine logstore.

METRIC_RAFT_ENGINE_APPEND_BATCH_ELAPSED

Timer of the append_batch operation on the raft-engine logstore.

METRIC_RAFT_ENGINE_READ_BYTES_TOTAL

Counter of bytes of the read operation on the raft-engine logstore.

METRIC_RAFT_ENGINE_READ_CALLS_TOTAL

Counter of calls of the read operation on the raft-engine logstore.

METRIC_RAFT_ENGINE_READ_ELAPSED

Timer of the append_batch operation on the raft-engine logstore. This timer only measures the duration of the read operation, not measures the total duration of replay.

Constants

LOGSTORE_LABEL

Logstore label.

OPTYPE_LABEL

Operation type label.